Allow comments/headers in the partition input
For example the Infomap clustering algorithm (see [0]) can produce a
clustering file that contains a header in the form "*Vertices 5000",
where 5000 is the number of vertices. It would make sense to read this
number but in order to keep the partition reader simple it is simply
ignored for now.

[0]: http://mapequation.org/code.html

--HG--
branch : Dev