Replaced the exact diameter calculation by the iFub-algorithm
This replaces the exact diameter calculation by the iFub-algorithm which
is in the worst case as slow as the old diameter algorithm but on most
real world networks it is actually very fast, sometimes even faster than
the estimated diameter.

This diameter calculation now also works for graphs that are not
connected, in this case the diameter of the connected component with the
largest diameter is returned.

--HG--
branch : Dev