Fix NMI calculation by not using HashingOverlapper
This fixes the NMI calculation by using the correct overlap values and
not the ones of the HashingOverlapper which are incorrect especially for
larger numbers of clusters. This should also speed up the NMI
calculation. Without this change, the newly added test case crashed
NetworKit because of failed assertions.

--HG--
branch : Dev