Improved Havel-Hakimi generator for unrealizable degree sequences
While previously a separate check needed to be executed, unrealizable
degree sequences are now detected during generation. If wanted, the
graph can be generated with a modified degree sequence even if the
original degree sequence is unrealizable.

The test if the sequence is realizable has been adapted to not to
require a sorted degree sequence. Therefore, the degree sequence is also
no longer sorted in the Havel-Hakami generator which means that the
degree sequence is now actually the requested degree sequence and not
just a permutation.

--HG--
branch : Dev