Made the LFR generator modular and more strict concerning community sizes
It is now possible to set arbitrary degree sequences, internal degree
distributions and community size distributions or even already assign
the nodes to communities in beforehand. However it is strictly checked
if the given community size sequence or the given partition is
realizable, if not an exception is thrown.

--HG--
branch : Dev