Make recursion in PLM optional
There are papers which claim that the louvain method is better without
the recursion. For example "[...] we are using the lowest
hierarchical level, the one with the initial aggregations of nodes into
clusters. Using the lowest hierarchical level of the method allows us to
circumvent the well-known resolution limit of modularity optimization,
which would otherwise compromise its performance."
http://journals.aps.org/pre/abstract/10.1103/PhysRevE.89.049902#fulltext

--HG--
branch : Dev