Rename ConfigurationModelGenerator to EdgeSwitchingMarkovChainGenerator
The "configuration model" only considers multigraphs, therefore it does
not really make sense to name this generator
ConfigurationModelGenerator. In the literature, this generator has
different names but Edge-Switching Markov-Chain generator seems to be
the best description in my opinion (as I consider markov-chain
monte-carlo algorithm to be less descriptive).

--HG--
branch : Dev