Improved version of TriangleEdgeScore, deprecate ChibaNishizekiTriangleEdgeScore
The new version of TriangleEdgeScore does not contain that many
optimizations that avoid work (in theory) but therefore also needs less
checks which seems to be faster in practice. Furthermore no locks or
atomics are needed anymore. The performance seems to be much better
than before and in particular it is now similar to
ChibaNishizekiTriangleEdgeScore even when parallelism is disabled.
Therfore TriangleEdgeScore should be used instead of
ChibaNishizekiTriangleEdgeScore.

--HG--
branch : Dev