PrefixJaccardScore: massive performance improvement
In first experiments this seems to be at least 10 times faster. The code
now uses much less malloc calls by not using an edge vector for each
node (but one for all edges) and by not using std::set for calculating
neighborhood intersections. The set is replaced by a boolean vector for
each thread.

--HG--
branch : Dev